+++ CHANGELOG.md
@@ -11,6 +11,7 @@ New Rules (and new "comments" ruleset)
  - #387: **JavadocConsecutiveEmptyLines** rule (comments) - Checks for javadoc comments with more than one consecutive empty line.
  - #389: **JavadocEmptySeeTag** rule (comments) - Checks for empty @see tags within javadoc.
  - #389: **JavadocEmptyParamTag** rule (comments) - Checks for empty @param tags within javadoc.
+ - #389: **JavadocEmptyReturnTag** rule (comments) - Checks for empty @return tags within javadoc.
 
 Updated/Enhanced Rules and Bug Fixes
  - #388: **ClassJavadoc** rule: Moved from the "formatting" ruleset into the new "comments" ruleset. [BREAKING CHANGE]